MAJOR CLEANUP!
* rss_fetch
got rid of the options array, replaced it with a more PHP-like
solution of using defines.  constants are setup, with defaults, in the
function init()

got rid of the idiom of passing back an array, its was awkward to deal
with in PHP, and unusual (and consquently confusing to people).  now i
return true/false values, and try to setup error string where
appropiate (rss_cache has the most complete example of this)

change the logic for interacting with the cache

* rss_cache
major re-working of how error are handled.  tried to make the code
more resillient.  the cache is now much more aware of MAX_AGE, where
before this was being driven out of rss_fetch (which was silly)

* rss_parse
properly handles xml parse errors.  used to sail along blithely unaware.